server-ai/requirements.txt
fastapi==0.115.0
uvicorn==0.32.0
python-multipart==0.0.20
pillow==11.0.0
ultralytics==8.3.0
opencv-python==4.10.0.84
numpy==2.0.2
